Extract semantics utilities unto separate files
